starting to implement conv
