+++ layers.h
@@ -36,4 +36,4 @@ Layer *create_dense_layer_l(int input_size, int output_size, Activation activati
 
 // conv
 
-Layer *create_conv2d_layer_l(int stride, int padding, int kernel_size, int input_size, int output_size, Activation activation, float epsilon, float decay_rate);
+Layer *create_conv2d_layer_l(int stride, int padding, int kernel_size, int input_size, Activation activation, float epsilon, float decay_rate);
