+++ layers/conv2d.c
@@ -5,18 +5,16 @@
 #include "conv2d.h"
 #include "../matrix.h"
 
-Conv2DLayer *create_conv2d_layer(int stride, int padding, int kernel_size, int input_size, int output_size, Activation activation, float epsilon, float decay_rate) {
+Conv2DLayer *create_conv2d_layer(int stride, int padding, int kernel_size, int input_size, Activation activation, float epsilon, float decay_rate) {
     Conv2DLayer *layer = malloc(sizeof(Conv2DLayer));
     layer->stride = stride;
     layer->padding = padding;
     layer->kernel_size = kernel_size;
     layer->input_size = input_size;
-    layer->output_size = output_size;
     layer->activation = activation;
     layer->epsilon = epsilon;
     layer->decay_rate = decay_rate;
-    layer->weights = create_matrix(output_size, input_size);
-    layer->bias = create_matrix(output_size, 1);
+    // TODO calc ouput size using the formula from standford and create weights and bias from that
     return layer;
 }
 
@@ -37,7 +35,7 @@ void destroy_conv2d_layer(Conv2DLayer *layer) {
 }
 
 Matrix *forward_conv2d(Conv2DLayer *layer, Matrix *input) {
-    return input;
+    Matrix *padded_input = pad(input, layer->padding);
 }
 
 Matrix *backward_conv2d(Conv2DLayer *layer, Matrix *loss_gradient) {
