+++ layers/dense.c
@@ -20,19 +20,6 @@ DenseLayer *create_dense_layer(int input_size, int output_size, Activation activ
     return layer;
 }
 
-void initialize_dense_layer_xavier_norm(DenseLayer *layer) {
-    int input_size = layer->input_size;
-    int output_size = layer->output_size;
-
-    float lower_bound = -(sqrt(6) / sqrt(input_size + output_size));
-    float upper_bound = sqrt(6) / sqrt(input_size + output_size);
-    for (int j = 0; j < layer->weights->rows; j++) {
-        for (int k = 0; k < layer->weights->cols; k++) {
-            layer->weights->data[j][k] = (float) rand() / (float) (RAND_MAX / (upper_bound - lower_bound)) + lower_bound;
-        }
-    }
-}
-
 void destroy_dense_layer(DenseLayer *layer) {
     if (layer == NULL) {
         return;
